requirements.txt
gurobipy==10.0.1
glpk
pycddlib
polytope
pypoman
tabulate
matplotlib
numpy
scipy
ipyparallel
torchvision
plotly==5.14.1 
pandas
more-itertools
